+++ template/component/js/resources/nls/root/@component@-strings.js
@@ -5,9 +5,7 @@
 
 */
 define({
-  "root": {
-    "@component-name@" : {
-    		"sampleString": "The strings file can be used to manage translatable resources"
-    }
+  "@component-name@" : {
+    	"sampleString": "The strings file can be used to manage translatable resources"
   }
 });