+++ template/component/ts/resources/nls/root/@component@-strings.ts
@@ -1,0 +1,5 @@
+export = {
+  "@component-name@": {
+    "sampleString": "The strings file can be used to manage translatable resources"
+  }
+};