+++ template/component/tsx/resources/nls/@component@-strings.ts
@@ -1,7 +1,0 @@
-export = {
-  "root": {
-    "@component-name@" : {
-    		"sampleString": "The strings file can be used to manage translatable resources"
-    }
-  }
-};